:recycle: refactor: 프로모션 검증 여부는 조직 생성과 분리 응답하도록 변경 (#89)
